Update Highlighter and Highlight existing with serialize and deSerialize in textDocument, remove listeners when unclicked
